renamed DashboardForm to ProjectPage
